+++ packages/flutter_meedu/lib/navigator.dart
@@ -1,0 +1,4 @@
+library navigator;
+
+export 'src/navigation/navigator/navigator.dart';
+export 'src/navigation/navigator/navigator_mixin.dart';