+++ packages/flutter_meedu/lib/router.dart
@@ -1,6 +1,5 @@
 library router;
 
 export 'src/navigation/navigator_observer.dart';
-export 'src/navigation/router/navigator.dart';
 export 'src/navigation/router/router.dart';
 export 'src/navigation/router/utils.dart';